refactor : LottoNumbers 싱글톤
- LottoNumbers 싱글톤
